feature(crud-services):get, put and delete methods added
